Better code organisation

+++ TVShows/TVShows/Home/ShowModels.swift
@@ -4,7 +4,7 @@ struct Show: Decodable {
     let description: String?
     let imageUrl: String?
     let noOfReviews: Int?
-    let averageRating: Int?
+    let averageRating: Double?
 
     enum CodingKeys: String, CodingKey {
         case id
